Refactor the code to be more maintainable
The intention of this change is to, as the description says, make the
code more maintainable.

As it was, it did the job that was required. However, everything was all
dumped in the main function. From what I've been learning about
idiomatic Go (and have learned over the years writing software) that's a
bad thing to do.

So, this change reformats it, breaking the large main function up into
a series of structs and functions, ones that could be further
restructured in a custom module, if desired.